Changed LinkEngine back into library. Switched all gui over to EngineGUI Project.
Will begin seperating gui into its own framework
All libraries will be posted to their own GitHub pages for ease of viewing
