+++ EngineGUI/EngineGUI.csproj
@@ -6,7 +6,7 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{C3707B9B-1193-4EC2-988B-626480103B5B}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>EngineGUI</RootNamespace>
+    <RootNamespace>LinkEngine</RootNamespace>
     <AssemblyName>EngineGUI</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -91,6 +91,7 @@
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
     </Compile>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
