+++ EngineGUI/GUI/NewProjectWindow.cs
@@ -15,11 +15,17 @@ namespace LinkEngine
         public NewProjectWindow()
         {
             InitializeComponent();
+            txtFile.Text = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments) + "\\LinkEngine\\Projects";
         }
 
         private void btnSubmit_Click(object sender, EventArgs e)
         {
             Hide();
+        }
+        private void btnBrowse_Click(object sender, EventArgs e)
+        {
+            folderBrowserDialog1.SelectedPath = txtFile.Text;
+            folderBrowserDialog1.ShowDialog();
         }
     }
 }
