+++ LinkEngine/LinkEngine.csproj
@@ -5,7 +5,7 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{63CF1CF9-7D8E-4B8B-825D-442F16DBFAEF}</ProjectGuid>
-    <OutputType>WinExe</OutputType>
+    <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>LinkEngine</RootNamespace>
     <AssemblyName>LinkEngine</AssemblyName>
@@ -52,30 +52,6 @@
     <Compile Include="Engine\Components\Physics\Raycast.cs" />
     <Compile Include="Engine\Components\Physics\Size.cs" />
     <Compile Include="Engine\GameObject.cs" />
-    <Compile Include="GUI\GUI.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="GUI\GUI.Designer.cs">
-      <DependentUpon>GUI.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\NewProjectWindow.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="GUI\NewProjectWindow.Designer.cs">
-      <DependentUpon>NewProjectWindow.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\NewFileForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="GUI\NewFileForm.Designer.cs">
-      <DependentUpon>NewFileForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\MainMenu.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="GUI\MainMenu.Designer.cs">
-      <DependentUpon>MainMenu.cs</DependentUpon>
-    </Compile>
     <Compile Include="Engine\Rendering\Screen.cs" />
     <Compile Include="Engine\Components\Physics\Collider2D.cs" />
     <Compile Include="Engine\Components\Physics\Transform.cs" />
@@ -84,7 +60,6 @@
     <Compile Include="Engine\Entities\Enemy.cs" />
     <Compile Include="Engine\Entities\Entity.cs" />
     <Compile Include="Engine\Entities\Player.cs" />
-    <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Engine\Components\UI\HUDObject.cs" />
     <Compile Include="Engine\WorldGen\Modules\Biome.cs" />
@@ -96,20 +71,6 @@
     <Compile Include="Engine\WorldGen\Modules\Tile.cs" />
     <Compile Include="Engine\WorldGen\Modules\Transition.cs" />
     <Compile Include="Engine\WorldGen\World.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="GUI\GUI.resx">
-      <DependentUpon>GUI.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\NewProjectWindow.resx">
-      <DependentUpon>NewProjectWindow.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\NewFileForm.resx">
-      <DependentUpon>NewFileForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\MainMenu.resx">
-      <DependentUpon>MainMenu.cs</DependentUpon>
-    </EmbeddedResource>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
